finplatform/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
PyJWT
passlib[bcrypt]
pydantic[email]
httpx
pytest
hypothesis
pytest-asyncio
anyio[trio]
